[Oztechan/CCC#1779] Move BaseDBDataSource and its test into database module (#1780)
* [Oztechan/CCC#1779] Move BaseDBDataSource and its test into database module

* [Oztechan/CCC#1779] Move BaseDBDataSource and its test into database module

* [Oztechan/CCC#1779] Move BaseDBDataSource and its test into database module
