--- common/src/commonTest/kotlin/com/oztechan/ccc/common/datasource/BaseDataSourceTest.kt
@@ -1,43 +1,0 @@
-package com.oztechan.ccc.common.datasource
-
-import com.oztechan.ccc.common.core.database.error.DatabaseException
-import com.oztechan.ccc.common.core.database.sql.CurrencyQueries
-import com.oztechan.ccc.common.datasource.currency.CurrencyDataSource
-import com.oztechan.ccc.common.datasource.currency.CurrencyDataSourceImpl
-import com.oztechan.ccc.test.BaseSubjectTest
-import com.oztechan.ccc.test.util.createTestDispatcher
-import io.mockative.Mock
-import io.mockative.classOf
-import io.mockative.given
-import io.mockative.mock
-import kotlinx.coroutines.test.runTest
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertFailsWith
-import kotlin.test.assertNotNull
-
-@Suppress("OPT_IN_USAGE")
-class BaseDataSourceTest : BaseSubjectTest<CurrencyDataSource>() {
-    override val subject: CurrencyDataSource by lazy {
-        CurrencyDataSourceImpl(currencyQueries, createTestDispatcher())
-    }
-
-    @Mock
-    private val currencyQueries = mock(classOf<CurrencyQueries>())
-
-    @Test
-    fun `any exception returns DatabaseException`() = runTest {
-        val exception = Exception("Test exception")
-        given(currencyQueries)
-            .invocation { getActiveCurrencies() }
-            .thenThrow(exception)
-
-        assertFailsWith(DatabaseException::class) {
-            subject.getActiveCurrencies()
-        }.let {
-            assertNotNull(it.cause)
-            assertEquals(exception, it.cause)
-            assertEquals(exception.message, it.cause!!.message)
-        }
-    }
-}
